Update css util
